perf(app): оптимизация размера бандла
