+++ .qwen/QWEN.md
@@ -1,0 +1,2 @@
+## Qwen Added Memories
+- Пользователь работает с проектом web на Next.js 16, React 19, tRPC, ZenStack/Prisma (PostgreSQL), FSD архитектурой, pnpm пакетным менеджером. Важно: использовать gen:client для OpenAPI, db:generate для ZModel, тесты через vitest/playwright.
