+++ .qwen/settings.json
@@ -1,0 +1,6 @@
+{
+  "permissions": {
+    "allow": ["Bash(pnpm *)", "WebSearch", "WebFetch(github.com)", "WebFetch(nextjs.org)"]
+  },
+  "$version": 3
+}
