+++ sentry.server.config.ts
@@ -1,6 +1,7 @@
 import * as Sentry from "@sentry/nextjs";
 
-import { IS_PROD, SENTRY_DSN } from "@/shared/constants/env.client";
+import { SENTRY_DSN } from "@/shared/constants/env.client";
+import { IS_PROD } from "@/shared/constants/env.flags";
 
 Sentry.init({
   dsn: SENTRY_DSN,
