+++ src/proxy.ts
@@ -6,7 +6,8 @@ import { routing } from "./i18n/routing";
 import { logger } from "./server/shared/infrastructure/logger";
 import { redisClient } from "./server/shared/infrastructure/redis";
 import { generateRequestId, getIp, sanitizeRequestId } from "./server/shared/lib/request-context";
-import { API_PREFIX, IS_PROD } from "./shared/constants/env.client";
+import { API_PREFIX } from "./shared/constants/env.client";
+import { IS_PROD } from "./shared/constants/env.flags";
 import { TURNSTILE_SECRET_KEY } from "./shared/constants/env.server";
 import { LOCALE_REGEX_STR } from "./shared/constants/locales";
 import { getCookieName } from "./shared/lib/utils";
