+++ src/server/shared/infrastructure/posthog-server.ts
@@ -1,15 +1,15 @@
 import { PostHog } from "posthog-node";
 
-import { env } from "@/shared/constants/env";
+import { NEXT_PUBLIC_POSTHOG_HOST, NEXT_PUBLIC_POSTHOG_KEY } from "@/shared/constants/env.client";
 
 let posthogClient: null | PostHog = null;
 
 export function getPostHogClient(): PostHog {
   if (posthogClient == null) {
-    posthogClient = new PostHog(env.NEXT_PUBLIC_POSTHOG_KEY, {
+    posthogClient = new PostHog(NEXT_PUBLIC_POSTHOG_KEY, {
       flushAt: 1,
       flushInterval: 0,
-      host: env.NEXT_PUBLIC_POSTHOG_HOST,
+      host: NEXT_PUBLIC_POSTHOG_HOST,
     });
   }
   return posthogClient;
