+++ src/server/shared/infrastructure/realtime.ts
@@ -1,6 +1,6 @@
 import Ably from "ably";
 
-import { IS_PROD } from "@/shared/constants/env.client";
+import { IS_PROD } from "@/shared/constants/env.flags";
 import { ABLY_API_KEY } from "@/shared/constants/env.server";
 
 const globalForAbly = globalThis as unknown as { ably?: Ably.Rest };
