+++ src/shared/constants/env.client.ts
@@ -1,4 +1,40 @@
-import { env } from "./env";
+import { createEnv } from "@t3-oss/env-nextjs";
+import { z } from "zod/v4-mini";
+
+import { envShared, isSharedValidationSkipped, sharedSchema } from "./env.shared";
+
+export const envClient = createEnv({
+  client: {
+    NEXT_PUBLIC_API_PREFIX: z
+      .string()
+      .check(z.startsWith("/"), z.regex(/^\/[\w/\-]*$/, "Invalid prefix format")),
+    NEXT_PUBLIC_APP_URL: z.optional(z.url()),
+    NEXT_PUBLIC_POSTHOG_HOST: z.string().check(z.minLength(1)),
+    NEXT_PUBLIC_POSTHOG_KEY: z.string().check(z.minLength(1)),
+    NEXT_PUBLIC_SENTRY_DSN: z.url(),
+    NEXT_PUBLIC_TRPC_PREFIX: z
+      .string()
+      .check(z.startsWith("/"), z.regex(/^\/[\w/\-]*$/, "Invalid prefix format")),
+    NEXT_PUBLIC_TURNSTILE_SITE_KEY: z.string().check(z.minLength(1)),
+  },
+
+  emptyStringAsUndefined: true,
+
+  runtimeEnv: {
+    NEXT_PUBLIC_API_PREFIX: process.env.NEXT_PUBLIC_API_PREFIX,
+    NEXT_PUBLIC_APP_URL: process.env.NEXT_PUBLIC_APP_URL,
+    NEXT_PUBLIC_POSTHOG_HOST: process.env.NEXT_PUBLIC_POSTHOG_HOST,
+    NEXT_PUBLIC_POSTHOG_KEY: process.env.NEXT_PUBLIC_POSTHOG_KEY,
+    NEXT_PUBLIC_SENTRY_DSN: process.env.NEXT_PUBLIC_SENTRY_DSN,
+    NEXT_PUBLIC_TRPC_PREFIX: process.env.NEXT_PUBLIC_TRPC_PREFIX,
+    NEXT_PUBLIC_TURNSTILE_SITE_KEY: process.env.NEXT_PUBLIC_TURNSTILE_SITE_KEY,
+    ...envShared,
+  },
+
+  shared: sharedSchema,
+
+  skipValidation: isSharedValidationSkipped,
+});
 
 const normalizePrefix = (value: string | undefined): `/${string}` => {
   if (value == null) {
@@ -7,22 +43,17 @@ const normalizePrefix = (value: string | undefined): `/${string}` => {
   return (value.startsWith("/") ? value : `/${value}`) as `/${string}`;
 };
 
-export const NODE_ENV = env.NODE_ENV;
-export const PORT = env.PORT;
-export const API_PREFIX = normalizePrefix(env.NEXT_PUBLIC_API_PREFIX);
-export const TRPC_PREFIX = normalizePrefix(env.NEXT_PUBLIC_TRPC_PREFIX);
+export const NODE_ENV = envClient.NODE_ENV;
+export const PORT = envClient.PORT;
+export const API_PREFIX = normalizePrefix(envClient.NEXT_PUBLIC_API_PREFIX);
+export const TRPC_PREFIX = normalizePrefix(envClient.NEXT_PUBLIC_TRPC_PREFIX);
 
-export const TURNSTILE_SITE_KEY = env.NEXT_PUBLIC_TURNSTILE_SITE_KEY;
-
-export const IS_DEV = env.NODE_ENV === "development";
-export const IS_PROD = env.NODE_ENV === "production";
-export const IS_TEST = env.NODE_ENV === "test";
-export const IS_CI = env.CI === "true";
-export const IS_ANALYZE = env.ANALYZE === "true";
+export const TURNSTILE_SITE_KEY = envClient.NEXT_PUBLIC_TURNSTILE_SITE_KEY;
 
 export const APP_URL =
-  env.NEXT_PUBLIC_APP_URL ?? (IS_DEV ? "http://localhost:3000" : "https://doxynix.space");
+  envClient.NEXT_PUBLIC_APP_URL ??
+  (process.env.NODE_ENV === "development" ? "http://localhost:3000" : "https://doxynix.space");
 
-export const SENTRY_DSN = env.NEXT_PUBLIC_SENTRY_DSN;
-export const NEXT_PUBLIC_POSTHOG_KEY = env.NEXT_PUBLIC_POSTHOG_KEY;
-export const NEXT_PUBLIC_POSTHOG_HOST = env.NEXT_PUBLIC_POSTHOG_HOST;
+export const SENTRY_DSN = envClient.NEXT_PUBLIC_SENTRY_DSN;
+export const NEXT_PUBLIC_POSTHOG_KEY = envClient.NEXT_PUBLIC_POSTHOG_KEY;
+export const NEXT_PUBLIC_POSTHOG_HOST = envClient.NEXT_PUBLIC_POSTHOG_HOST;
