+++ src/shared/constants/env.server.ts
@@ -1,44 +1,107 @@
 import "server-only";
 
-import { env } from "./env";
+import { createEnv } from "@t3-oss/env-nextjs";
+import { z } from "zod/v4-mini";
 
-export const JWT_SECRET = env.JWT_SECRET;
-export const NEXTAUTH_SECRET = env.NEXTAUTH_SECRET;
-export const DATABASE_URL = env.DATABASE_URL;
-export const BETTERSTACK_TOKEN = env.BETTERSTACK_API_TOKEN;
-export const TURNSTILE_SECRET_KEY = env.TURNSTILE_SECRET_KEY;
-export const RESEND_API_KEY = env.RESEND_API_KEY;
-export const ABLY_API_KEY = env.ABLY_API_KEY;
-export const GROQ_API_KEY = env.GROQ_API_KEY;
-export const UPLOADTHING_TOKEN = env.UPLOADTHING_TOKEN;
-export const GITHUB_APP_ID = env.GITHUB_APP_ID;
-const rawGithubAppPrivateKey = env.GITHUB_APP_PRIVATE_KEY;
+import { envShared, isSharedValidationSkipped, sharedSchema } from "./env.shared";
 
-if (env.NODE_ENV !== "test" && typeof rawGithubAppPrivateKey !== "string") {
+export const envServer = createEnv({
+  runtimeEnv: {
+    ABLY_API_KEY: process.env.ABLY_API_KEY,
+    BETTERSTACK_API_TOKEN: process.env.BETTERSTACK_API_TOKEN,
+    DATABASE_URL: process.env.DATABASE_URL,
+    GITHUB_APP_ID: process.env.GITHUB_APP_ID,
+    GITHUB_APP_PRIVATE_KEY: process.env.GITHUB_APP_PRIVATE_KEY,
+    GITHUB_CLIENT_ID: process.env.GITHUB_CLIENT_ID,
+    GITHUB_CLIENT_SECRET: process.env.GITHUB_CLIENT_SECRET,
+    GITHUB_SYSTEM_INSTALLATION_ID: process.env.GITHUB_SYSTEM_INSTALLATION_ID,
+    GITHUB_SYSTEM_PAT: process.env.GITHUB_SYSTEM_PAT,
+    GITHUB_WEBHOOK_SECRET: process.env.GITHUB_WEBHOOK_SECRET,
+    GITLAB_CLIENT_ID: process.env.GITLAB_CLIENT_ID,
+    GITLAB_CLIENT_SECRET: process.env.GITLAB_CLIENT_SECRET,
+    GOOGLE_CLIENT_ID: process.env.GOOGLE_CLIENT_ID,
+    GOOGLE_CLIENT_SECRET: process.env.GOOGLE_CLIENT_SECRET,
+    GROQ_API_KEY: process.env.GROQ_API_KEY,
+    JWT_SECRET: process.env.JWT_SECRET,
+    NEXTAUTH_SECRET: process.env.NEXTAUTH_SECRET,
+    RESEND_API_KEY: process.env.RESEND_API_KEY,
+    TURNSTILE_SECRET_KEY: process.env.TURNSTILE_SECRET_KEY,
+    UPLOADTHING_TOKEN: process.env.UPLOADTHING_TOKEN,
+    YANDEX_CLIENT_ID: process.env.YANDEX_CLIENT_ID,
+    YANDEX_CLIENT_SECRET: process.env.YANDEX_CLIENT_SECRET,
+    ...envShared,
+  },
+
+  server: {
+    ABLY_API_KEY: z.string().check(z.minLength(1)),
+    BETTERSTACK_API_TOKEN: z.string().check(z.minLength(1)),
+    DATABASE_URL: z.url(),
+
+    GITHUB_APP_ID: z.string().check(z.regex(/^\d+$/, "must be numeric")),
+    GITHUB_APP_PRIVATE_KEY: z
+      .string()
+      .check(z.regex(/BEGIN .*PRIVATE KEY/), z.regex(/END .*PRIVATE KEY/)),
+    GITHUB_CLIENT_ID: z.string().check(z.minLength(1)),
+    GITHUB_CLIENT_SECRET: z.string().check(z.minLength(1)),
+    GITHUB_SYSTEM_INSTALLATION_ID: z.string().check(z.regex(/^\d+$/)),
+    GITHUB_SYSTEM_PAT: z.optional(z.string().check(z.minLength(1))),
+    GITHUB_WEBHOOK_SECRET: z.string().check(z.minLength(1)),
+    GITLAB_CLIENT_ID: z.string().check(z.minLength(1)),
+    GITLAB_CLIENT_SECRET: z.string().check(z.minLength(1)),
+    GOOGLE_CLIENT_ID: z.string().check(z.minLength(1)),
+    GOOGLE_CLIENT_SECRET: z.string().check(z.minLength(1)),
+    GROQ_API_KEY: z.string().check(z.minLength(1)),
+    JWT_SECRET: z.string().check(z.minLength(1)),
+    NEXTAUTH_SECRET: z.string().check(z.minLength(1)),
+    RESEND_API_KEY: z.string().check(z.minLength(1)),
+    TURNSTILE_SECRET_KEY: z.string().check(z.minLength(1)),
+    UPLOADTHING_TOKEN: z.string().check(z.minLength(1)),
+    YANDEX_CLIENT_ID: z.string().check(z.minLength(1)),
+    YANDEX_CLIENT_SECRET: z.string().check(z.minLength(1)),
+  },
+
+  shared: sharedSchema,
+
+  skipValidation: isSharedValidationSkipped,
+});
+
+export const JWT_SECRET = envServer.JWT_SECRET;
+export const NEXTAUTH_SECRET = envServer.NEXTAUTH_SECRET;
+export const DATABASE_URL = envServer.DATABASE_URL;
+export const BETTERSTACK_TOKEN = envServer.BETTERSTACK_API_TOKEN;
+export const TURNSTILE_SECRET_KEY = envServer.TURNSTILE_SECRET_KEY;
+export const RESEND_API_KEY = envServer.RESEND_API_KEY;
+export const ABLY_API_KEY = envServer.ABLY_API_KEY;
+export const GROQ_API_KEY = envServer.GROQ_API_KEY;
+export const UPLOADTHING_TOKEN = envServer.UPLOADTHING_TOKEN;
+export const GITHUB_APP_ID = envServer.GITHUB_APP_ID;
+const rawGithubAppPrivateKey = envServer.GITHUB_APP_PRIVATE_KEY;
+
+if (envServer.NODE_ENV !== "test" && typeof rawGithubAppPrivateKey !== "string") {
   throw new Error("GITHUB_APP_PRIVATE_KEY is required and must be a valid PEM string");
 }
 
 export const GITHUB_APP_PRIVATE_KEY =
   typeof rawGithubAppPrivateKey === "string" ? rawGithubAppPrivateKey.replaceAll("\\n", "\n") : "";
-export const GITHUB_SYSTEM_PAT = env.GITHUB_SYSTEM_PAT ?? null;
-export const GITHUB_WEBHOOK_SECRET = env.GITHUB_WEBHOOK_SECRET;
-export const GITHUB_SYSTEM_INSTALLATION_ID = env.GITHUB_SYSTEM_INSTALLATION_ID;
+export const GITHUB_SYSTEM_PAT = envServer.GITHUB_SYSTEM_PAT ?? null;
+export const GITHUB_WEBHOOK_SECRET = envServer.GITHUB_WEBHOOK_SECRET;
+export const GITHUB_SYSTEM_INSTALLATION_ID = envServer.GITHUB_SYSTEM_INSTALLATION_ID;
 
 export const AUTH_PROVIDERS = {
   github: {
-    id: env.GITHUB_CLIENT_ID,
-    secret: env.GITHUB_CLIENT_SECRET,
+    id: envServer.GITHUB_CLIENT_ID,
+    secret: envServer.GITHUB_CLIENT_SECRET,
   },
   gitlab: {
-    id: env.GITLAB_CLIENT_ID,
-    secret: env.GITLAB_CLIENT_SECRET,
+    id: envServer.GITLAB_CLIENT_ID,
+    secret: envServer.GITLAB_CLIENT_SECRET,
   },
   google: {
-    id: env.GOOGLE_CLIENT_ID,
-    secret: env.GOOGLE_CLIENT_SECRET,
+    id: envServer.GOOGLE_CLIENT_ID,
+    secret: envServer.GOOGLE_CLIENT_SECRET,
   },
   yandex: {
-    id: env.YANDEX_CLIENT_ID,
-    secret: env.YANDEX_CLIENT_SECRET,
+    id: envServer.YANDEX_CLIENT_ID,
+    secret: envServer.YANDEX_CLIENT_SECRET,
   },
 };
