+++ src/tests/integration/property.test.ts
@@ -3,7 +3,7 @@ import { enhance } from "@zenstackhq/runtime";
 import * as fc from "fast-check";
 import { afterAll, beforeAll, beforeEach, describe, it, vi } from "vitest";
 
-import { IS_CI } from "@/shared/constants/env.client";
+import { IS_CI } from "@/shared/constants/env.flags";
 
 import { prisma } from "@/server/shared/infrastructure/db";
